+++ lib/generators/lite/ruby/templates/install.rb
@@ -2,6 +2,6 @@
 
 Lite::Ruby.configure do |config|
   config.monkey_patches = %w[
-    array date enumerable hash integer kernel numeric object range string struct time
+    array date enumerable hash integer kernel numeric object open_struct range string struct time
   ]
 end