+++ lib/lite/ruby/struct.rb
@@ -2,6 +2,14 @@
 
 class Struct
 
+  def [](key)
+    send(key)
+  end
+
+  def []=(key, val)
+    send("#{key}=", val)
+  end
+
   def attributes
     each_pair.with_object({}) { |(key, val), hash| hash[key] = val }
   end